(PUP-2860) Remove repeated assert_rgen from other sources.
The assert_rgen_installed() method is called from multiple sources,
and it did not check if an assert had already been made. This logic
was in the methods that called assert_rgen_installed().

This change moves the check if already asserted into the method
itself. This reduces the number of calls to require rgen, and the number
of rgen version compliance checks to 1.