(PUP-2506) Trivial Ruby bug in Puppet::Resource::Ral
Prior to this commit, a bug is triggered when the Puppet::Error string
is interpolated. The #type method was called recursively and failed with
an ArgumentError: wrong number of arguments (0 for 1).

This commit fixes the bug by changing which variable is interpolated in
the string.  It also adds a spec test for this case.