(maint) Add sitemoduledir to env env tests
These two tests set up a 'special' environment, and made the assumption
that the modulepath could be reset to something that did not include the
sitemoduledir.  In PE, not having this directory on the modulepath will
case a pluginsync on the master to remove all of the PE modules,
potentially causing problems for subsequent runs.  Adding the
sitemoduledir to the modulepath if we are running a pe master ensures
the pluginsync libdir is not cleared.