(PUP-514) Make Puppet::Resource behave as PResourceType wrt access
This makes Puppet::Resource behave as a PResourceType when using
the access operator; to detail a type with a title, or to
access a parameter of a typed/titled resource.

This is done to make differences between 3x and future planned catalog
builder transparent. 4x uses PResourceType as a reference to a
resource (handle/future ref), where 3x uses instances of
Puppet::Resource for both the instance, and references to it.