(maint) Remove unnecessary setting of modulepath from face help spec
The integration spec checking help output for the faces was setting
modulepath directly in the body of an RSpec context block.  This caused
modulepath to be set on test file load.  Normally, this is not an issue,
since most specs are not sensitive to the Puppet modulepath setting, and
the test_helper clears all settings in its after_each_test hook.
However, a recent run of specs on Solaris master just happened to run
with the very first spec being sensitive to modulepath
(spec/integration/directory_environments_spec.rb) which then failed.
Fortunately, the current version of puppet no longer requires this early
setting of modulepath for the documentation spec to run, so I'm just
removing it here.