(PUP-2583) Fix bad test case in file type mode acceptance test.
The test case was assuming that +X on a file would always set the
requested executable bit.  This is only true if the resource is a
directory and not a file.  For files, there must be an already set
executable bit in the permission for +X to have an effect.

There was a bug in Puppet that was causing this acceptance test to pass
with the incorrect assumption.  That bug was fixed in commit ac2622e8c1
and now the acceptance test needs to be updated for the correct behavior.