(PUP-514) Fix assignability of 0 min Tuples
The assignability logic didn't consider an empty Tuple (Tuple) to be assignable
to a Tuple of min size 0 (Tuple[Object, 0, default]). This prevented the
case of a Callable that represented with only varargs from being called
with no arguments (Callable[Object, 0, default], Tuple).

The mistake was that after the check that the sizes of the tuples were
compatible, it then used the *max* of the two tuple's type parameters to
determine if the two are compatible. However, only the list of types in
the right hand side tuple matter in determining assignability.