(PUP-1782) Add deprecation for regexp match against non string
This adds deprecation to use of the match operators and the LHS
is not a string.
This also adds a deprecation warning when a match is made with
a regexp in a case or selector and a match is found - this because
these will not match in the future, but those that do not match
will continue to not match.