(PUP-2737) fix an error in the user type unit test for ssh key purging
After fixing PUP-2737, it became apparent that there was an error in the
expectations agains the fixture for the user type unit tests. A key
from the fixture did contain spaces in its comment, but the unit test
actually failed to expect the correct value.

Now that the full resource title is retrieved, make a slight adjustment
so that the test keeps working and uses a more appropriate name for the
key.