(PUP-514) Extract block evaluation into evaluator
The code for handling closure calls was mostly in the evaluator, but in
order to get the type checking correct, the split was becoming difficult
to deal with. This changes it so that closures are implemented in the
compiler as evaluation of blocks in a new scope with a given set of
variable bindings. This moves all of the argument handling into the
closure class, which will combine better with type checking.