Remove Build depends on facter << 2
Lucid is failing to build because it cannot satisfy facter << 2.0. This was
originally to prevent it from pulling in the erroneous facter 2, but there's no
reason for it as a build depends. Now that we've shipped facter 2 to our lucid
repos, even though lucid upstream has facter 1.5.6, we're getting build
failures because of facter.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>