(PUP-121) Ensure classes are absolute when transforming to Resource
When the 3x to 4x calls take place, resource references in the form
of a PCatalogEntryType are transformed to Puppet::Resource.
Depending on where these (reference) Puppet::Resource instances are
used they need to either be absolute, or not (yes, 3x is inconsistent).

Since future parser in 3.7 (and later 4.0) will not support dynamic
name scoping, the PCatalogEntryType's should in general be absolute.

This commit changes the transformation to make the created
Puppet::Resource (references) absolute (when it is possible).

This also adds more tests