(PUP-514) Extract method to match args with parameters
The call() method of Closure was getting unwieldy. This extracts one of
the large portions of it and starts working it down into something a
little bit smaller.

Because this is on the normal path for evaluation, I don't want to
extract too many methods. Every method call is significant overhead here
:(