(PUP-2666) Unskip neg tests for PMT generate interview
Change criteria of negative tests in order to allow them to
execute.

The scope of the negative tests has been changed to only evaluating
that the proper error response is given when unsupported input is
given for a module version.

With this rescoping, these tests no longer need to be skipped.