(maint) Stop using ambiguous Manager name
Manager is not a very good name for a class. What does it manage? What
does it mean to manage that? How does it manage? Who manages the
manager? This renames the Profiler::Manager class to just
Profiler::AroundProfiler because it modifies profilers to work so that
they automatically go around a block. This is then the interface exposed
by the Puppet::Util::Profiler singleton.