(PUP-514) Invert use of parameters_specified
When see_scope was renamed to parameters_specified it was done
incorrectly and actually meant parameters_not_specified. The logic
seemed wrong and convoluted at the time, but I couldn't see how. Joshua
Partlow pointed out that parameters_specified was being set to true when
the parameters were actually *not* specified. By flipping all of this
around it all becomes a bit more obvious.