(PUP-121) Fix dynamic environments test to accept ::absolute classnames
One of the dynamic environment tests is an attempt to include a class
that is not in the default manifests during a puppet apply.  This
produces an expected error about the missing class; however, PUP-121
changed the message slightly because for future parser, relative
classnames get transformed to be absolute and an attempt to 'include
testing_mod' ends up reporting that ::testing_mod could not be found.
This commit just updates the test to check for ::testing_mod not being
found when working with future parser.