(PUP-2755) Mark function stubs as rvalue
The function stubs that are left in place to allow "puppet doc" to
document the new functions were supposed to issue an error when they are
used under "--evaluate current", but you couldn't call them in their
normal fashion because they were not specified as :rvalue functions,
which they used to be. This puts the rvalue flag back on so that if they
are called then they can produce the intended error message.