(PUP-2781) Generate correct dependency expressions for new modules.
Since Puppet 3.6.0, `puppet module generate` has automatically created
a `metadata.json` file for users, with a sample dependency on
`puppetlabs-stdlib` to help demonstrate the syntax for expressing
dependencies.  Unfortunately, the generated sample used an incorrect
key name (`version_range` vs `version_requirement`), which is less
useful and was introducing other problems with (e.g.) publishing to the
Forge.

This change repairs that mistake.