(PUP-2828) Pin to 1.9.3 branch of vendored ruby on Windows
Previously, acceptance tests on Windows would install the default branch
of our vendored ruby repo. This is a bit fragile.

This commit makes it explicit that we are using the `1.9.3` branch,
which contains the older set of windows gems.

Newer development to support x64 will be done in the `1.9.3-x86` branch.