(maint) Use an absolute path so we're not sensitive to pwd
Previously, the location of the windows versions.yaml file depended on
your current working directory, and assumed it was the root of the
project. But acceptance tests are typically executed by cd'ing into
acceptance, and then executing rake tasks from there.

This commit ensures we reference the versions.yaml relative to the
currently executing 000_EnvSetup.rb setup step.