(maint) Extract constant for name separator
The name separator is used a few different times in the file. This
extracts it out and freezes the value, which has the nice benefit of a
small speed boost. Unfortunately this isn't a big win for the system
since this code path is only really used for error conditions.