(PUP-2104) Fix facter terminus to print out fact files.
Prior to the refactoring of the facter terminus, the terminus loaded
module facts directly.  This is no longer the case and instead facter
always loads the facts itself.

The previous implementation logged an informational message displaying
the path to each fact script it found.  The new behavior only printed
the directory that would be searched for facts.

Given that this information may be useful to module developers and to
maintain existing output behavior, this change restores the previous
output behavior but still relies on facter to load the facts.